Allow series data to be appended for live updates
When an AJAX request is made, there are now two ways that it is handled.
If the graph does not yet exist, then the request will overwirte any
existing series data with the new data payload and the onComplete method
is called (both as before). However, if the graph already exists, the
request will *append* the new data to the relevant series and the
onComplete method is *not* called. This allows efficient live updates of
existing graphs when new data arrive.